remove test where createFotter is included
