+++ packages/shared/src/settings/securityGlobalControl.ts
@@ -2,6 +2,7 @@
 
 import { SettingBlock, createFooter } from '@frontify/guideline-blocks-settings';
 import { Security } from './types';
+import { getSecurityGlobalControlId } from './securityDownloadable';
 
 /**
  * Returns segment controls for global security settings.
@@ -38,33 +39,4 @@ export const getSecurityGlobalControlSetting = (id?: string): SettingBlock[] => 
             }),
         },
     ];
-};
-
-/**
- * Returns the downloadable security settings.
- *
- * @param {string} id custom id for the setting block
- * @param {string} globalControlId custom id for the global control setting block
- *
- * @returns {SettingBlock} Returns downloadable security settings.
- */
-
-type SecurityDownloadableSettingType = {
-    id?: string;
-    globalControlId?: string;
-};
-
-export const getSecurityDownloadableSetting = (options?: SecurityDownloadableSettingType): SettingBlock => {
-    const securityId = getSecurityGlobalControlId(options?.globalControlId);
-    return {
-        id: options?.id ? options.id : 'downloadable',
-        type: 'switch',
-        defaultValue: false,
-        label: 'Downloadable',
-        show: (bundle) => bundle.getBlock(securityId)?.value === Security.Custom,
-    };
-};
-
-const getSecurityGlobalControlId = (id?: string): string => {
-    return id ? id : 'security';
 };
